Change seed from int64 to uint64

+++ vms/proposervm/proposer/windower.go
@@ -113,7 +113,7 @@ func (w *windower) Proposers(ctx context.Context, chainHeight, pChainHeight uint
 	}
 
 	seed := chainHeight ^ w.chainSource
-	w.sampler.Seed(int64(seed))
+	w.sampler.Seed(seed)
 
 	indices, err := w.sampler.Sample(numToSample)
 	if err != nil {
